Add low level wheel control
The WheelXXX classes are used to control a rotation direction and speed
of a single wheel. The wheel can be driven by a Continues Rotation Servo
or be a Brushed DC Motor driven via an HBRidge.

The Wheel object abstracts the rotation control such that the actual
drive mechanism is irrelevant.

Speed is managed as a percentage of full speed, i.e. a speed value of
100% is full speed in the forward direction, 0% is stopped and -100% is
full speed in the reverse direction.

Forward and reverse mechanical rotation is determined by the hardware
connections, pin assignments (in case of HBridge) and side of the robot
the wheel is attached to.

Rotation direction can also be reversed in software.