First test of Cybot firmware
This is the first test of combining the various modules into one sketch
to create the Cybot firmware.

The DriveTrain was changed to make it a Task which will check for any new
commands that it needs to action, and issue the required changes to the
Wheels.

Some minor comment changes is to start bringing Class, method and
function comments into the headers files instead of the code files. The
argument is that more often I would look in the header file for usage
information than in the code file. We'll see how this goes.