Setup up software serial port (SSerial) for Bluetooth
The Bluetooth input method used to directly connect to, and set up, it's
own software serial port for input only. Since this port was only
available inside the BT input handler, it was not easy to use the BT
port for general output too, without hooking this into the BT input
handler.

By creating a generic Software Serial port, called SSerial, set up and
instantiated via the config.{h,cpp} files based on config and enabling
defines in config.h, the BT input handler can use this for input, and
any other parts of the system can use SSerial for output.

The SSerial port then behaves exactly like the built in hardware Serial
and we also create an `OpenSoftSerial()` helper for it in utils, like
the `OpenSerial()` helper for Serial, and also a debug output macros
called DSS like we have the D macro for Serial.

TODO: Refactoring of the SerialIn and BTIn objects because there is a
lot of duplication of code there now.