WIP: crash bug to be found and fixed still
The code crashes the mcu and I'm not sure where the problem is yet.
Seems it may be related to memory, so just comitting what I have so far
on this branch so I can go back to a working version and start doing
some memory profiling to get an idea of memory usage on the working
version.

Will then come bcak here to do the same profiling and see if it changes
much in this version.

Another possibility is that it has to do with the analog sensor readings
being taken for the LDR. Need to explore this if memory is not an issue.